pyproject.toml
[project]
name = "financial-agent"
version = "0.1.0"
description = "Assistente financeiro em português brasileiro usando LangGraph"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Arthur Granja", email = "your.email@example.com" }
]
license = { file = "LICENSE" }
keywords = ["financial", "agent", "langchain", "langgraph", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-dotenv>=1.0.1",
    "langchain>=0.3.12",
    "langgraph>=0.2.55",
    "langgraph-checkpoint-sqlite>=2.0.3",
    "langchain-openai>=0.2.10",
    "langchain-anthropic>=0.2.8",
    "langchain-google-genai>=2.0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
financial-chat = "scripts.chat:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
